Fixed video and genre tests

+++ .docker/entrypoint.sh
@@ -1,8 +1,14 @@
 #!/bin/bash
 
 #On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
-cp .env.example .env
-cp .env.testing.example .env.testing
+if [ ! -f ".env" ]; then
+    cp .env.example .env
+fi
+
+if [ ! -f ".env.testing" ]; then
+    cp .env.testing.example .env.testing
+fi
+
 composer install
 php artisan key:generate
 php artisan migrate
